Move components up a folder in tree
