+++ src/components/Error.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
-import { ErrorProps } from '../../interfaces/index';
-import { httpResponses } from '../../utils/httpResponses'
+import { ErrorProps } from '../interfaces/index';
+import { httpResponses } from '../utils/httpResponses'
 
 export const Error: React.FC<ErrorProps> = ({ errorCode, clearSelected }) => {
   let matchingResponse = httpResponses.find(response =>
