+++ src/index.tsx
@@ -1,7 +1,7 @@
 import ReactDOM from 'react-dom';
 import { HashRouter } from 'react-router-dom';
 import './index.css';
-import { App } from './components/app/App';
+import { App } from './components/App';
 
 const router = <HashRouter> <App /> </HashRouter>;
